+++ Java/src/main/java/com/gildedrose/Items/GildedRoseItem.java
@@ -1,4 +1,4 @@
-package com.gildedrose;
+package com.gildedrose.Items;
 
 public interface GildedRoseItem {
     public void updateQuality();
