+++ Java/src/main/java/com/gildedrose/Items/Item.java
@@ -1,4 +1,4 @@
-package com.gildedrose;
+package com.gildedrose.Items;
 
 public class Item {
