+++ Java/src/main/java/com/gildedrose/Items/LegendaryItem.java
