+++ Java/src/main/java/com/gildedrose/Items/NormalItem.java
