+++ Java/src/test/java/com/gildedrose/TexttestFixture.java
@@ -1,5 +1,7 @@
 package com.gildedrose;
 
+import com.gildedrose.Items.Item;
+
 public class TexttestFixture {
     public static void main(String[] args) {
         System.out.println("OMGHAI!");
